fix: some changes with resolve bug and code smell
