+++ src/apollo/vote.ts
@@ -2,6 +2,23 @@ import { CHAINS, ChainId, GovernanceType } from '@pangolindex/sdk';
 import gql from 'graphql-tag'; // eslint-disable-line import/no-named-as-default
 import { getGovernanceSubgraphApolloClient } from 'src/apollo/client';
 
+export type NFT_PROPOSAL = {
+  id: string;
+  targets: string[];
+  values: string[];
+  signatures: string[];
+  calldatas: string[];
+  proposer?: string;
+  forVotes: number;
+  againstVotes: number;
+  startTime: number;
+  endTime: number;
+  eta: any;
+  description: string;
+  executed: boolean;
+  canceled: boolean;
+};
+
 export const GET_PROPOSALS = gql`
   query proposals($where: Proposal_filter) {
     proposals(orderBy: startTime, orderDirection: desc, where: $where) {
@@ -22,7 +39,7 @@ export const GET_PROPOSALS = gql`
   }
 `;
 
-export const GET_HEDERA_PROPOSALS = gql`
+export const GET_NFT_GOVERNANCE_PROPOSALS = gql`
   query proposals($where: Proposal_filter) {
     proposals(orderBy: startTime, orderDirection: desc, where: $where) {
       id
@@ -48,12 +65,14 @@ export const getAllProposalData = async (chainId: ChainId, id?: string) => {
     return null;
   }
 
-  let data = [] as Array<any>;
+  let data = [] as Array<NFT_PROPOSAL>;
 
   try {
     const queryData: any = {
       query:
-        CHAINS[chainId]?.contracts?.governor?.type === GovernanceType.SAR_NFT ? GET_HEDERA_PROPOSALS : GET_PROPOSALS, // TODO Condition
+        CHAINS[chainId]?.contracts?.governor?.type === GovernanceType.SAR_NFT
+          ? GET_NFT_GOVERNANCE_PROPOSALS
+          : GET_PROPOSALS,
       fetchPolicy: 'cache-first',
     };
