+++ src/state/governance/hooks/common.ts
@@ -1,9 +1,16 @@
 import { JSBI } from '@pangolindex/sdk';
 import { ethers, utils } from 'ethers';
-import { useEffect, useState } from 'react';
-import { getAllProposalData } from 'src/apollo/vote';
+import { useEffect, useMemo, useState } from 'react';
+import { NFT_PROPOSAL, getAllProposalData } from 'src/apollo/vote';
+import { PROPOSAL_STORAGE_INTERFACE } from 'src/constants/governance/proposalStorage';
 import { useChainId } from 'src/hooks';
+import { useSarNftGovernanceAssistantContract, useSarNftGovernanceContract } from 'src/hooks/useContract';
 import { useAllProposalDataHook } from 'src/state/governance/hooks';
+import {
+  useMultipleContractSingleData,
+  useSingleCallResult,
+  useSingleContractMultipleData,
+} from 'src/state/pmulticall/hooks';
 import { ProposalData, ProposalState } from '../types';
 
 export const enumerateProposalState = (state: number) => {
@@ -11,7 +18,7 @@ export const enumerateProposalState = (state: number) => {
   return proposalStates[state];
 };
 
-export const getProposalState = (proposal: ProposalData) => {
+export const getProposalState = (proposal: NFT_PROPOSAL) => {
   const currentTimestamp = () => Math.floor(Date.now() / 1000);
 
   if (proposal.canceled) {
@@ -21,8 +28,8 @@ export const getProposalState = (proposal: ProposalData) => {
   } else if (currentTimestamp() <= proposal.endTime) {
     return ProposalState.active;
   } else if (
-    proposal.againstCount &&
-    JSBI.lessThanOrEqual(JSBI.BigInt(proposal.forCount), JSBI.BigInt(proposal?.againstCount || 0))
+    proposal.againstVotes &&
+    JSBI.lessThanOrEqual(JSBI.BigInt(proposal.forVotes), JSBI.BigInt(proposal?.againstVotes || 0))
   ) {
     return ProposalState.defeated;
   } else if (proposal.eta === 0) {
@@ -68,9 +75,7 @@ export function useGetProposalsViaSubgraph(id?: string) {
             title: proposal?.description?.split(/# |\n/g)[1] || 'Untitled',
             description: proposal?.description || 'No description.',
             proposer: proposal?.proposer,
-            status:
-              getProposalState({ ...proposal, forCount: proposal?.forVotes, againstCount: proposal?.againstVotes }) ??
-              'Undetermined',
+            status: getProposalState({ ...proposal }) ?? 'Undetermined',
             forCount: proposal?.forVotes ? parseFloat(ethers.utils.formatUnits(proposal?.forVotes.toString(), 18)) : 0,
             againstCount: proposal?.againstVotes
               ? parseFloat(ethers.utils.formatUnits(proposal?.againstVotes.toString(), 18))
@@ -103,4 +108,99 @@ export function useProposalData(id: string): ProposalData | undefined {
 
   const allProposalData = useAllProposalData();
   return allProposalData?.find((p) => p.id === id);
+}
+
+// get count of all proposals made
+export function useProposalCount(): number | undefined {
+  const gov = useSarNftGovernanceContract();
+  const res = useSingleCallResult(gov, 'proposalCount');
+  if (res.result && !res.loading) {
+    return parseInt(res.result[0]) ?? 0;
+  }
+  return undefined;
+}
+
+// get data for all past and active proposals
+export function useSarNftAllProposalData() {
+  const proposalCount = useProposalCount();
+  const govContract = useSarNftGovernanceContract();
+  const govAssistantContract = useSarNftGovernanceAssistantContract();
+
+  const proposalIndexes = [] as any;
+  for (let i = 1; i <= (proposalCount ?? 0); i++) {
+    proposalIndexes.push([i]);
+  }
+
+  // get all proposal entities
+  const allProposalsAddressesState = useSingleContractMultipleData(
+    govAssistantContract,
+    'locateProposal',
+    proposalIndexes,
+  );
+
+  // get the address of the rewarder for each pool
+  const allProposalsAddresses = useMemo(() => {
+    if ((allProposalsAddressesState || []).length === 0) return [];
+    return allProposalsAddressesState.map((proposal) => {
+      if (!!proposal?.result?.address) {
+        return proposal?.result?.address;
+      }
+      return undefined;
+    });
+  }, [allProposalsAddressesState]);
+
+  const allProposals = useMultipleContractSingleData(
+    allProposalsAddresses,
+    PROPOSAL_STORAGE_INTERFACE,
+    'getProposal',
+    [],
+  );
+
+  // get all proposal states
+  const allProposalStates = useSingleContractMultipleData(govContract, 'state', proposalIndexes);
+
+  if (allProposals && allProposalStates) {
+    allProposals.reverse();
+    allProposalStates.reverse();
+
+    return allProposals
+      .filter((p, i) => {
+        return Boolean(p.result) && Boolean(allProposalStates[i]?.result);
+      })
+      .map((_p, i) => {
+        const description = allProposals[i]?.result?.description;
+
+        const details = (allProposals[i]?.result?.targets || []).map((target: string, i: number) => {
+          const signature = allProposals[i]?.result?.signatures[i];
+
+          const [name, types] = signature?.substr(0, signature?.length - 1)?.split('(') || [];
+
+          const calldata = allProposals[i]?.result?.calldatas[i];
+
+          const decoded = utils.defaultAbiCoder.decode(types.split(','), calldata);
+
+          return {
+            target,
+            functionSig: name,
+            callData: decoded.join(', '),
+          };
+        });
+
+        const formattedProposal: ProposalData = {
+          id: allProposals[i]?.result?.id.toString(),
+          title: description?.split(/# |\n/g)[1] || 'Untitled',
+          description: description || 'No description.',
+          status: enumerateProposalState(allProposalStates[i]?.result?.[0]) ?? 'Undetermined',
+          forCount: parseFloat(ethers.utils.formatUnits(allProposals[i]?.result?.forVotes.toString(), 18)),
+          againstCount: parseFloat(ethers.utils.formatUnits(allProposals[i]?.result?.againstVotes.toString(), 18)),
+          startTime: parseInt(allProposals[i]?.result?.startTime?.toString()),
+          endTime: parseInt(allProposals[i]?.result?.endTime?.toString()),
+          startBlock: parseInt(allProposals[i]?.result?.startBlock?.toString()),
+          details: details,
+        };
+        return formattedProposal;
+      });
+  } else {
+    return [];
+  }
 }
